move the test case to org.apache.ibatis.autoconstructor
